Actualizacion: Correciones menores
