+++ .gitignore
@@ -15,6 +15,7 @@ chrome-profiler-events*.json
 speed-measure-plugin*.json
 
 # IDEs and editors
+.vscode
 /.idea
 .project
 .classpath
