--- src/app/components/stepper-example/stepper-example.component.css
@@ -1,230 +1,0 @@
-* {
-	margin: 0;
-	padding: 0;
-}
-
-html {
-	height: 100%;
-}
-
-p {
-	color: grey;
-}
-
-#heading {
-	text-transform: uppercase;
-	color: #ff3a28; /** #673ab7; **/
-	font-weight: normal;
-}
-
-#msform {
-	text-align: center;
-	position: relative;
-	margin-top: 20px;
-}
-
-#msform fieldset {
-	background: white;
-	border: 0 none;
-	border-radius: 0.5rem;
-	box-sizing: border-box;
-	width: 100%;
-	margin: 0;
-	padding-bottom: 20px;
-	position: relative;
-}
-
-.form-card {
-	text-align: left;
-}
-
-#msform fieldset:not(:first-of-type) {
-	display: none;
-}
-
-#msform input[type='text'],
-#msform textarea,
-#msform select {
-	padding: 8px 15px 8px 15px;
-	border: 1px solid #ccc;
-	border-radius: 0px;
-	margin-bottom: 15px;
-	margin-top: 2px;
-	width: 100%;
-	box-sizing: border-box;
-	font-family: montserrat;
-	color: #2c3e50;
-	background-color: #eceff1;
-	font-size: 16px;
-	letter-spacing: 1px;
-}
-
-#msform input:focus,
-#msform textarea:focus,
-#msform select:focus {
-	-moz-box-shadow: none !important;
-	-webkit-box-shadow: none !important;
-	box-shadow: none !important;
-	border: 1px solid #ff3a28;
-	outline-width: 0;
-}
-
-#msform select option {
-	background-color: #eceff1;
-}
-
-#msform .action-button {
-	width: 100px;
-	background: #ff3a28;
-	font-weight: bold;
-	color: white;
-	border: 0 none;
-	border-radius: 0px;
-	cursor: pointer;
-	padding: 10px 5px;
-	margin: 10px 0px 10px 5px;
-	float: right;
-}
-
-#msform .action-button:hover,
-#msform .action-button:focus {
-	background-color: #df1526;
-}
-
-#msform .action-button-previous {
-	width: 100px;
-	background: #616161;
-	font-weight: bold;
-	color: white;
-	border: 0 none;
-	border-radius: 0px;
-	cursor: pointer;
-	padding: 10px 5px;
-	margin: 10px 5px 10px 0px;
-	float: right;
-}
-
-#msform .action-button-previous:hover,
-#msform .action-button-previous:focus {
-	background-color: #000000;
-}
-
-.card {
-	z-index: 0;
-	border: none;
-	position: relative;
-}
-
-.fs-title {
-	font-size: 25px;
-	color: #ff3a28;
-	margin-bottom: 15px;
-	font-weight: normal;
-	text-align: left;
-}
-
-.purple-text {
-	color: #ff3a28;
-	font-weight: normal;
-}
-
-.steps {
-	font-size: 25px;
-	color: gray;
-	margin-bottom: 10px;
-	font-weight: normal;
-	text-align: right;
-}
-
-.fieldlabels {
-	color: gray;
-	text-align: left;
-}
-.danger {
-	color: #df1526;
-	text-align: right;
-}
-
-#progressbar {
-	margin-bottom: 30px;
-	overflow: hidden;
-	color: lightgrey;
-}
-
-#progressbar .active {
-	color: #ff3a28;
-}
-
-#progressbar li {
-	list-style-type: none;
-	font-size: 15px;
-	width: 25%;
-	float: left;
-	position: relative;
-	font-weight: 400;
-}
-
-#progressbar #account:before {
-	font-family: FontAwesome;
-	content: '\f13e';
-}
-
-#progressbar #personal:before {
-	font-family: FontAwesome;
-	content: '\f007';
-}
-
-#progressbar #payment:before {
-	font-family: FontAwesome;
-	content: '\f030';
-}
-
-#progressbar #confirm:before {
-	font-family: FontAwesome;
-	content: '\f00c';
-}
-
-#progressbar li:before {
-	width: 50px;
-	height: 50px;
-	line-height: 45px;
-	display: block;
-	font-size: 20px;
-	color: #ffffff;
-	background: lightgray;
-	border-radius: 50%;
-	margin: 0 auto 10px auto;
-	padding: 2px;
-}
-
-#progressbar li:after {
-	content: '';
-	width: 100%;
-	height: 2px;
-	background: lightgray;
-	position: absolute;
-	left: 0;
-	top: 25px;
-	z-index: -1;
-}
-
-#progressbar li.active:before,
-#progressbar li.active:after {
-	background: #ff3a28;
-}
-
-.progress {
-	height: 20px;
-}
-
-.progress-bar {
-	background-color: #ff3a28;
-}
-
-.fit-image {
-	width: 100%;
-	object-fit: cover;
-}
-.list-group {
-	margin-bottom: 25px;
-}
