+++ src/app/services/rut.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { FormControl, FormGroup } from '@angular/forms';
+import { FormControl } from '@angular/forms';
 import { validate, clean, format } from 'rut.js';
 
 interface ErrorValidate {
@@ -11,23 +11,20 @@ interface ErrorValidate {
 export class RutService {
 	constructor() {}
 
-	public rutValidate(control: FormControl): ErrorValidate {
+	public rutValido(control: FormControl): ErrorValidate {
 		if (!control.value) {
 			return null;
 		}
 
 		let cleanValue: string = clean(control.value);
-		console.log(cleanValue);
 
 		let esValido: boolean = validate(cleanValue);
 
-		console.log(`Es un rut valido? -> ${esValido}`);
 		if (esValido) {
 			let formateado: string = format(cleanValue);
-
 			control.setValue(formateado);
 			return {
-				correcto: true,
+				rutValido: true,
 			};
 		} else {
 			return null;
