+++ src/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
 	<head>
 		<meta charset="utf-8" />
-		<title>SamplingIzitapp</title>
+		<title>Formulario de venta</title>
 		<meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
 		<base href="/" />
 		<!--BOOTSTRAP FRAMEWORK-->
@@ -10,8 +10,6 @@
 		<link rel="stylesheet" href="assets/css/bootstrap/bootstrap-checkbox-radio-list-group-item.min.css" />
 		<!--FONTAWESOME-->
 		<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.0.3/css/font-awesome.css" />
-		<!--GOOGLE-->
-		<script src="https://maps.googleapis.com/maps/api/js?libraries=places&key=AIzaSyDqpl0ozSQ_YzgWHRQ8JhOqKYJRM0UjiKc"></script>
 	</head>
 	<body>
 		<app-root></app-root>
@@ -26,6 +24,6 @@
 		<script src="assets/js/main.js"></script>
 		 -->
 
-		<script src="assets/js/example.js"></script>
+		<script src="assets/js/stepper.js"></script>
 	</body>
 </html>
